app/requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
sqlalchemy
psycopg2-binary
faiss-cpu
langchain-experimental  # <-- Needed for the SQL agent
pypdf
pdfplumber            # <-- To read PDFs
pandas                # <-- To read CSVs
python-dotenv
tqdm
numpy==1.26.4
langgraph